@@ -1,22 +1,25 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
-
 import 'package:supercharged/supercharged.dart';
 
 class AlbumCard extends StatelessWidget {
+  const AlbumCard(
+    this._albumTitle,
+    this._albumArtist,
+    this._cardColor, {
+    Key key,
+  }) : super(key: key);
   final String _albumTitle;
   final String _albumArtist;
   final String _cardColor;
 
-  AlbumCard(this._albumTitle, this._albumArtist, this._cardColor);
-
   @override
   Widget build(BuildContext context) {
-    return Container(
+    return SizedBox(
       width: 360,
       child: Card(
         shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(15.0),
+          borderRadius: BorderRadius.circular(15),
         ),
         color: _cardColor.toColor(),
         elevation: 10,
@@ -24,32 +27,45 @@ class AlbumCard extends StatelessWidget {
           mainAxisSize: MainAxisSize.min,
           children: <Widget>[
             ListTile(
-              leading: Icon(Icons.album, size: 70.0),
-              title: Text(_albumTitle,
-                  style: GoogleFonts.poppins(
-                      fontSize: 20.0,
-                      fontWeight: FontWeight.w500,
-                      fontStyle: FontStyle.italic)),
-              subtitle: Text(_albumArtist.toUpperCase(),
-                  style: GoogleFonts.raleway(
-                      fontSize: 16.0, fontWeight: FontWeight.w700)),
+              leading: const Icon(Icons.album, size: 70),
+              title: Text(
+                _albumTitle,
+                style: GoogleFonts.poppins(
+                  fontSize: 20,
+                  fontWeight: FontWeight.w500,
+                  fontStyle: FontStyle.italic,
+                ),
+              ),
+              subtitle: Text(
+                _albumArtist.toUpperCase(),
+                style: GoogleFonts.raleway(
+                  fontSize: 16,
+                  fontWeight: FontWeight.w700,
+                ),
+              ),
             ),
             ButtonBarTheme(
-              data: ButtonBarThemeData(),
+              data: const ButtonBarThemeData(),
               child: ButtonBar(
                 children: <Widget>[
-                  FlatButton(
-                    textColor: '#d8dee9'.toColor(),
-                    child: Text('Songs'.toUpperCase(),
-                        style: GoogleFonts.raleway(
-                            fontSize: 16.0, fontWeight: FontWeight.w700)),
+                  ElevatedButton(
+                    child: Text(
+                      'Songs'.toUpperCase(),
+                      style: GoogleFonts.raleway(
+                        fontSize: 16,
+                        fontWeight: FontWeight.w700,
+                      ),
+                    ),
                     onPressed: () {},
                   ),
-                  FlatButton(
-                    textColor: '#eceff4'.toColor(),
-                    child: Text('Shop'.toUpperCase(),
-                        style: GoogleFonts.raleway(
-                            fontSize: 16.0, fontWeight: FontWeight.w700)),
+                  ElevatedButton(
+                    child: Text(
+                      'Shop'.toUpperCase(),
+                      style: GoogleFonts.raleway(
+                        fontSize: 16,
+                        fontWeight: FontWeight.w700,
+                      ),
+                    ),
                     onPressed: () {},
                   ),
                 ],
