@@ -1,12 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
 import 'package:google_fonts/google_fonts.dart';
-
 import 'package:kenburns/kenburns.dart';
 import 'package:supercharged/supercharged.dart';
 
-import 'package:fl_shop/widgets/album_card.dart';
+import 'widgets/album_card.dart';
 
 void main() {
   SystemChrome.setSystemUIOverlayStyle(
@@ -14,71 +12,86 @@ void main() {
       statusBarColor: '#fffcfa'.toColor(),
     ),
   );
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
+  const MyApp({Key key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
-    return new MaterialApp(
-        debugShowCheckedModeBanner: false,
-        theme: ThemeData(
-            primaryColor: '#fffcfa'.toColor(),
-            textTheme: TextTheme(
-              body1: GoogleFonts.poppins(textStyle: TextStyle(fontSize: 16.0)),
-            )),
-        home: NewsApp());
+    return MaterialApp(
+      debugShowCheckedModeBanner: false,
+      theme: ThemeData(
+        colorSchemeSeed: Colors.brown[600],
+        useMaterial3: true,
+      ),
+      home: const NewsApp(),
+    );
   }
 }
 
 class NewsApp extends StatelessWidget {
+  const NewsApp({Key key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       body: DefaultTabController(
         length: 2,
         child: NestedScrollView(
-          headerSliverBuilder: (BuildContext context, bool innerBoxIsScrolled) {
+          headerSliverBuilder: (context, innerBoxIsScrolled) {
             return <Widget>[
               SliverOverlapAbsorber(
-                  handle:
-                      NestedScrollView.sliverOverlapAbsorberHandleFor(context),
-                  sliver: SliverAppBar(
-                    expandedHeight: 300.0,
-                    elevation: 0.0,
-                    floating: false,
-                    pinned: true,
-                    flexibleSpace: FlexibleSpaceBar(
-                        centerTitle: true,
-                        title: Text("Let's shop",
-                            style: GoogleFonts.raleway(
-                                textStyle: TextStyle(
-                              color: '#2e3440'.toColor(),
-                              fontSize: 16.0,
-                              fontWeight: FontWeight.w700,
-                            ))),
-                        background: KenBurns(
-                            child: Image.asset(
-                              'assets/undraw_add_to_cart_vkjp.png',
-                              fit: BoxFit.cover,
-                            ))),
-                  )),
+                handle:
+                    NestedScrollView.sliverOverlapAbsorberHandleFor(context),
+                sliver: SliverAppBar(
+                  expandedHeight: 300,
+                  elevation: 0,
+                  pinned: true,
+                  flexibleSpace: FlexibleSpaceBar(
+                    centerTitle: true,
+                    title: Text(
+                      "Let's shop",
+                      style: GoogleFonts.raleway(
+                        textStyle: TextStyle(
+                          color: '#2e3440'.toColor(),
+                          fontSize: 16,
+                          fontWeight: FontWeight.w700,
+                        ),
+                      ),
+                    ),
+                    background: KenBurns(
+                      child: Image.asset(
+                        'assets/undraw_add_to_cart_vkjp.png',
+                        fit: BoxFit.cover,
+                      ),
+                    ),
+                  ),
+                ),
+              ),
               SliverPersistentHeader(
                 delegate: _SliverAppBarDelegate(
                   TabBar(
                     indicatorColor: '#bf616a'.toColor(),
-                    indicatorWeight: 5.0,
+                    indicatorWeight: 5,
                     indicatorPadding:
-                        EdgeInsets.symmetric(vertical: 0.0, horizontal: 16.0),
+                        const EdgeInsets.symmetric(horizontal: 16),
                     labelColor: '#2e3440'.toColor(),
                     labelStyle: GoogleFonts.cabin(
-                        textStyle: TextStyle(
-                            fontSize: 16.0, fontWeight: FontWeight.w500)),
+                      textStyle: const TextStyle(
+                        fontSize: 16,
+                        fontWeight: FontWeight.w500,
+                      ),
+                    ),
                     unselectedLabelColor: '#4c566a'.toColor(),
                     unselectedLabelStyle: GoogleFonts.cabin(
-                        textStyle: TextStyle(
-                            fontSize: 16.0, fontWeight: FontWeight.w400)),
-                    tabs: [
+                      textStyle: const TextStyle(
+                        fontSize: 16,
+                        fontWeight: FontWeight.w400,
+                      ),
+                    ),
+                    tabs: const [
                       Tab(icon: Icon(Icons.shop), text: 'Shop'),
                       Tab(icon: Icon(Icons.shopping_cart), text: 'Cart'),
                     ],
@@ -86,24 +99,24 @@ class NewsApp extends StatelessWidget {
                   '#fffcfa'.toColor().withOpacity(0.97),
                 ),
                 pinned: true,
-                floating: false,
               ),
             ];
           },
-          body: ListView(children: <Widget>[
-            Column(
-              mainAxisAlignment: MainAxisAlignment.start,
-              children: <Widget>[
-                Padding(padding: EdgeInsets.fromLTRB(4.0, 72.0, 4.0, 0.0)),
-                AlbumCard('Dark & Wild', 'BTS', '#d08770'),
-                Padding(padding: EdgeInsets.fromLTRB(4.0, 12.0, 4.0, 0.0)),
-                AlbumCard('Wings', 'BTS', '#a3beac'),
-                Padding(padding: EdgeInsets.fromLTRB(4.0, 12.0, 4.0, 0.0)),
-                AlbumCard('Love Yourself: Tear', 'BTS', '#b48ead'),
-                Padding(padding: EdgeInsets.fromLTRB(4.0, 12.0, 4.0, 0.0)),
-              ],
-            )
-          ]),
+          body: ListView(
+            children: <Widget>[
+              Column(
+                children: const <Widget>[
+                  Padding(padding: EdgeInsets.fromLTRB(4, 72, 4, 0)),
+                  AlbumCard('Dark & Wild', 'BTS', '#d08770'),
+                  Padding(padding: EdgeInsets.fromLTRB(4, 12, 4, 0)),
+                  AlbumCard('Wings', 'BTS', '#a3beac'),
+                  Padding(padding: EdgeInsets.fromLTRB(4, 12, 4, 0)),
+                  AlbumCard('Love Yourself: Tear', 'BTS', '#b48ead'),
+                  Padding(padding: EdgeInsets.fromLTRB(4, 12, 4, 0)),
+                ],
+              )
+            ],
+          ),
         ),
       ),
     );
@@ -123,10 +136,13 @@ class _SliverAppBarDelegate extends SliverPersistentHeaderDelegate {
 
   @override
   Widget build(
-      BuildContext context, double shrinkOffset, bool overlapsContent) {
-    return new Container(
-      child: _tabBar,
+    BuildContext context,
+    double shrinkOffset,
+    bool overlapsContent,
+  ) {
+    return ColoredBox(
       color: _color,
+      child: _tabBar,
     );
   }
 
